+++ src/DrawWrapper.cpp
@@ -1,6 +1,6 @@
 #include <DrawWrapper.hpp>
 
-void GUI::_draw_translate(int x, int y)
+void GUI::_draw_translate(float x, float y)
 {
     if (x == 0 && y == 0)
         return;
