+++ src/Root.cpp
@@ -25,7 +25,7 @@ void GUI::Root::draw_textShadow(int row, int column, char *text, Color color)
 {
     GUI::draw_textShadow(row, column, text, color);
 }
-void GUI::Root::draw_translate(int x, int y)
+void GUI::Root::draw_translate(float x, float y)
 {
     GUI::_draw_translate(x, y);
 }
