+++ src/Scroll.cpp
@@ -69,7 +69,7 @@ void GUI::Scroll::draw_textShadow(int row, int column, char *text, Color color)
 {
     super.draw_textShadow(row, column, text, color);
 }
-void GUI::Scroll::draw_translate(int x, int y)
+void GUI::Scroll::draw_translate(float x, float y)
 {
     super.draw_translate(x, y);
 }
