+++ src/WidgetBase.cpp
@@ -31,7 +31,7 @@ void GUI::WidgetBase::draw_textShadow(int row, int column, char *text, Color col
 {
     super.draw_textShadow(row, column, text, color);
 }
-void GUI::WidgetBase::draw_translate(int x, int y)
+void GUI::WidgetBase::draw_translate(float x, float y)
 {
     super.draw_translate(x, y);
 }
